+++ src/components/Toast/Toast.test.tsx
@@ -1,11 +1,9 @@
 import '@testing-library/jest-dom'
 
 import { emitToastToggle } from '@events/toast'
-import { fireEvent, render, screen, waitFor } from '@testing-library/react'
-import React from 'react'
-import { act } from 'react-dom/test-utils'
+import { act, fireEvent, render, screen, waitFor } from '@testing-library/react'
 
-import Toast from './Toast'
+import { Toast } from './Toast'
 
 jest.setTimeout(10000)
 
@@ -19,7 +17,12 @@ describe('Toast tests', () => {
       expect(screen.getByTestId('toast-component')).toBeInTheDocument()
     )
 
-    await act(async () => new Promise(r => setTimeout(r, 5001)))
+    await act(
+      async () =>
+        new Promise(r => {
+          setTimeout(r, 5001)
+        })
+    )
 
     expect(screen.queryByTestId('toast-component')).not.toBeInTheDocument()
   })
